+++ Eyes.Sdk.Core.DotNet/Server/BatchClose.cs
@@ -39,6 +39,12 @@ namespace Applitools
             return this;
         }
 
+        public BatchClose SetProxy(ProxySettings proxy)
+        {
+            WebProxy webProxy = new WebProxy(proxy.ProxyUri);
+            return SetProxy(webProxy);
+        }
+
         public BatchClose SetProxy(WebProxy proxy)
         {
             ArgumentGuard.NotNull(proxy, nameof(proxy));